requirements.txt
PyYAML~=6.0.1
beautifulsoup4~=4.12.2
Markdown~=3.4.4